Add the ability to complete full runs up to Paladin.
By default, the bot is in test mode. In this mode, it will simply enable
sequences if you are in the appropriate location and finish battles off as
appropriate. On the other hand, if you disable test mode and enable the
script, it will reset the emulator and automatically send splits to
LiveSplit at the conclusion of most battles.

LiveSplit integration requires luasocket to be installed into the BizHawk
directory. Instructions on accomplishing this should eventually be added
to a README, but it only requires copying a few files to certain
locations.

Signed-off-by: Jason Lynch <jason@calindora.com>