Make menu.field.item.select more robust.
Previously, it was pressing A at times where it wasn't guaranteed it would
work. The input delay masked this in most cases, but it reared its ugly
head in a couple of runs. This change may or may not fix it, depending on
if the chosen memory address use is consistent.

Signed-off-by: Jason Lynch <jason@calindora.com>