Move Big Whale boarding into the walk.board function.
The Big Whale is unique in that boarding it doesn't immediately change
your vehicle. Rather, it takes you to another map (the interior of the
ship), wherein you can ultimately activate the vehicle.

Using walk.interact to board the ship worked well, except when the bot
happened to run into an encounter on the Big Whale tile, which would cause
the interaction to be lost to oblivion. We fix this by allowing walk.board
to consider both vehicle changes and a change to the Big Whale interior
map.

Signed-off-by: Jason Lynch <jason@calindora.com>