Change the way the run operates and enables test mode.
Previously, the user had to set or unset a variable to enable test mode.
Now, the script enters test mode if the emulator frame counter is anything
other than 1. This ensures runs only begin with a fresh emulator state (by
loading the script, and then rebooting the emulator core). This should
also reduce the likelihood of desyncs and keep things consistent.

Additionally, the script expects to begin with an empty SRAM.

Signed-off-by: Jason Lynch <jason@calindora.com>