Add some basic configuration variables to the main script.
SEED, by default, is set to nil, which will cause the bot to do random
runs. Set this to a specific number to do a specific run. Note that runs
will only be reproducible with the same version of Edge.

LIVESPLIT, disabled by default, allows the user to enable or disable
LiveSplit integration.

Signed-off-by: Jason Lynch <jason@calindora.com>