Make opening the field item use dialog more robust.
If an encounter occurred on the tile where A would be pressed, simply
interacting may cause A to be pressed at a time when it won't do anything.
Since interact doesn't verify the result of its action in any way, the bot
could get stuck. This function instead verifies that the item use menu is
opened.

Signed-off-by: Jason Lynch <jason@calindora.com>