Delay Edge's initial action in the CPU battle.
Depending on the levels of the characters, the exact ATB changes. Without
this delay, Edge would sometimes act too quickly, resulting in the
Defender undoing some of Edge's initial damage. This is bad. The delay
solves the timing in this case and doesn't seem to adversely affect other
cases. If any adverse effect is discovered, it may be necessary to make
the delay level-specific.

Signed-off-by: Jason Lynch <jason@calindora.com>