Move the inventory sort to just after the grind fight.
The inventory sort is a stopgap until some form of proper inventory
management is implemented. Previously, it occurred immediately after
receiving the Protect ring, but as it turned out, sometimes the inventory
was full when attempting to pick up the ring. Prevent this by sorting
earlier, directly after the grind fight.

Signed-off-by: Jason Lynch <jason@calindora.com>