Increase the delay before entering Baron Castle.
This is probably an excessive delay, but the bot occasionally moved too
early with the lower delay. A proper solution would be to detect the
landing of the airship, wait a few frames, and then move. The current
delay is measured from when A is pressed to land the airship.

Signed-off-by: Jason Lynch <jason@calindora.com>