Avoid taking the post-battle spoils after the grind fight.
Instead of writing special code to handle the spoils dialog, we simply
take advantage of the fact that it presses the current mash button. We
temporarily set it to B so the screen will be canceled.

This is incompatible with eventually using the screen to get rid of items
if needed, however.

Signed-off-by: Jason Lynch <jason@calindora.com>