Improve the post-grind fight inventory management.
First, we only attempt to move Excalbur and Elixir if they're not in the
desired final location. Attempting to move them if they're already in the
correct place leads to bad things.

Second, in the case of having too few free slots, instead of sorting the
inventory, we simply drop a few items that are no longer needed. I'm not
sure this case will ever actually show up, but it's there for safety.

Signed-off-by: Jason Lynch <jason@calindora.com>