Tweak the grind fight recovery strategies slightly.
The recent change caused FuSoYa to queue up weak (while in the heal phase)
even if a dragon isn't currently detected, as long as someone recently
attacked the searcher. (This is probably going to eventually lead to a
softlock, if the attacker misses.) However, because the dragon was not yet
there, the variable that keeps track of the dragon's current HP was set to
0. As such, later characters would believe no damage has yet been done to
the dragon, and would not attack. This commit bandages over this error.

Signed-off-by: Jason Lynch <jason@calindora.com>