Fix a major cause of errant auto-reloading.
After dying, the game begins loading the game loading screen. While the cursor
is over a given save, the data from that save is actually loading into memory as
if that is the current state. Since the save game in question is right after the
Mist Clip, it corresponds to the beginning of the "Girl" sequence, which allows
for automatic reloading. We fix this by preventing it from activating sequences
in this situation.

Signed-off-by: Jason Lynch <jason@calindora.com>