Make the grind fight potentially slightly safer.
If the character that attacks the dragon misses, the bot will get stuck in a
loop of doing nothing but healing. This valiant effort eventually falls to the
dragon's Fire attack. Ideally, we'd just detect that the character's attack
missed, but until that functionality is available, we work around it by ensuring
that the same character will at least attack when their turn comes around.
Hopefully they don't miss again.

Signed-off-by: Jason Lynch <jason@calindora.com>