Fix and improve LiveSplit integration with the newest BizHawk.
We now provide basic instructions on how to set up the integration, though it
remains a difficult task for the time being.

This commit also enhances the LiveSplit integration to hopefully eliminate any
syncing problems it experienced in earlier versions (as in seeds behaving
differently depending on whether or not the integration was enabled).
Previously, if the script was unable to succesfully send the message to
LiveSplit, it would cause a delay of at least one frame, which would obviously
lead to sync problems. The new version operates asynchronously, queueing any
commands sent to it, and sending them to LiveSplit as it is able.

Signed-off-by: Jason Lynch <jason@calindora.com>