Sort the inventory if there are few empty spots and no TrashCan.
This situation never seems to happen to human runners, so it must be an artifact
of some sort of the way the bot manages its inventory. Nonetheless, if the bot
gets to near the end of the game and doesn't have at least three spots open, it
needs to clear some space. Previously, we simply dumped some items in the
TrashCan, but with the switch to TrashCan strats in the Milon Z fight, this is
no longer always a valid solution. In these cases, sort the inventory instead.

Signed-off-by: Jason Lynch <jason@calindora.com>