Add a safety delay when selecting characters in the item menu.
This appears to correct a problem with healing on at least one seed. However, I
am unsure if it actually fixes the issue or if it simply prevented it from
occurring in this particular instance.

In either case, it's apparent that a redesign of the menu function contracts is
in order.

Signed-off-by: Jason Lynch <jason@calindora.com>