Improve safety of the Magus Sisters battle.
First, we enhance the recovery options so both Cecil and Yang can revive
Tellah if he is dead, and change Cecil to be a little more forthcoming with
curing himself and Tellah.

Second, we ensure that Tellah is not muted prior to the battle by making that a
standard status restored automatically by the _restore_party function. We may
want to ultimately make _restore_party more versatile regarding statuses, but
I'm not sure how important that really is.

Signed-off-by: Jason Lynch <jason@calindora.com>