Experimental fix for the dragon's Fire occurring between Life/Elixir.
This fix attempts to wait for Fire before using the Life item. This may not work
as desired, as there may not be enough time to both queue and execute both the
Life and Elixir before the next fire. If this is the case, we may try a
different synchronization method.

Signed-off-by: Jason Lynch <jason@calindora.com>