Add additional code to handle potentially better handle a fully inventory.
This makes use of the fact that the inventory will be sorted during the Eblan
battle by sorting the inventory just prior, and then forcing the bot to resort
it to where we want things during the battle. It may be overkill, so hopefully
we only trigger it in dire inventory situations.

Signed-off-by: Jason Lynch <jason@calindora.com>