Fix a bug with the dialog code during reset.
Note that this doesn't actually make any sense to me, since it seems to stem
from a variable carrying over between runs. I figured reboot_core would
successfully clear the environments, but apparently not. If the bug continues to
occur we can investigate further. (Specifically, during the next run, the bot is
unable to set seed 92 because the dialog code is intensely mashing A.)

Signed-off-by: Jason Lynch <jason@calindora.com>