Add a mode to reset for time.
This code currently automatically resets after a split if it's a certain amount
behind the programmed splits (values from a recent PB, though not the absolute
PB I have ever seen). Approximately, it allows itself to be up to 2.5 minutes
behind at the beginning, narrowing the window to 60 seconds by the end. This
could probably be tuned or at least parameterized.

Signed-off-by: Jason Lynch <jason@calindora.com>