Implement a new inventory sorting algorithm.
The previous algorithm looked for errors in priority sequencing and made
adjustments to compensate. It would look for the highest error and then move
that item as far down as it thought it could go, but this would result in it
spending a lot of time doing small movements toward the top, while ignoring
some larger movements that could be made moving items toward the bottom.

The new algorithm sorts from the bottom up, locating (hopefully) the item to
move that will make the most impact. Testing has been minimal, however.

Signed-off-by: Jason Lynch <jason@calindora.com>