Implement an adjusted TrashCan strat for Milon Z and eliminate the Cure2 strat.
This version of the TrashCan strat is based on my own attempts. The primary
difference is that it unconditionally uses an item on Cecil's first turn,
quickly parries without a run buffer if he gets a second, and then proceeds as
usual. This hasn't yet failed me in a run that I recall, but obviously the
bot's mileage may vary. As the bot does more runs, the fight can be adjusted
again if necessary to avoid any pitfalls.

Signed-off-by: Jason Lynch <jason@calindora.com>