Update for the new Dwarf Axe-aware step routes.
Note that these are not fully tested. The routes themselves are still
beta and will probably change over the next few weeks. The no64-rosa
route is almost certainly broken (as some additional changes need to be
made). The no64-excalbur route appears to be working, though it may have
yet undiscovered issues.

Signed-off-by: Jason Lynch <jason@calindora.com>