Hopefully improve the reliability of menus opening.
This is ultimately intended to fix the problem of needing to immediately
menu after the FlameDog battle. For one frame, the current method is
tricked into thinking the menu is open, so it moves on to selecting the
desired command. This fix simply requires the command selectors to
verify that the menu is open before doing anything.

Signed-off-by: Jason Lynch <jason@calindora.com>