Initial version of the updated no64-rosa route.
This uses a new safer version of the Zeromus strat which involves Rydia.
This route is not entirely complete, and there are certain things that
can currently cause it to break down, especially Rydia not reaching
level 26 after CPU. There are also some minor issues with some of the
battles, especially regarding inventory management.

Signed-off-by: Jason Lynch <jason@calindora.com>