Attempt to fix the issue with the post-Elements menu.
The sequence previously executed before the Elements fight, so it would
be unable to determine Cecil's weapon after the fight. This commit
shifts things slightly so the CPU sequence begins after the fight. I
vaguely remember this section being problematic to get working correctly
early in development, so hopefully this doesn't break anything.

Signed-off-by: Jason Lynch <jason@calindora.com>