Fix some step route issues with the Hummingway Cave in the no64-excalbur route.
First, there is a chance that the player will be unable to get to the
shop and back to the entrance within the step limit. Therefore, it must
be prepared to cast Warp on occasion. In addition, there is the
possibility of stepping on two tiles that do not count as steps, leading
to an odd number of steps.

The current code is not foolproof, however. There are certain areas of
the room that will almost certainly result in a softlock. Whether it's
possible to actually get to any of them, however, is a different
question.

I have left in a few debug statements for a short while until I have
seen that the code does not fail frequently.

Signed-off-by: Jason Lynch <jason@calindora.com>