Attempt to continue using a step route in the event of a reset from a non-viable seed.
Seeds 74 through 81 will break the item duplication, so the script
resets and tries for a new seed. Unfortunately, this currently breaks
the step routing. It's still questionable whether it's valuable to have
these seeds for statistics gathering purposes, but in the case where you
truly want the bot to do random runs, it makes sense to have it do the
best it can even though it rolled a bad seed initially.

Signed-off-by: Jason Lynch <jason@calindora.com>